backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-SocketIO==5.3.5
Flask-Cors==4.0.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
openai
a2a-sdk
aioconsole
fastmcp
fqdn
google-adk
google-cloud-logging
google-generativeai
jupyterlab
langchain
networkx
sqlmodel
weave
webcolors
wikipedia
